requirements.txt
ace_tools==0.0
altgraph==0.17.4
asttokens==3.0.0
basyx-python-sdk @ git+https://github.com/eclipse-basyx/basyx-python-sdk.git@ecf4c13290ebe58cf99c6599d1f3963d68651eee
bidict==0.23.1
blinker==1.9.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==44.0.1
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
dnspython==2.6.1
et-xmlfile==1.1.0
executing==2.2.0
filelock==3.15.4
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
fonttools==4.53.0
fsspec==2024.6.1
h11==0.14.0
huggingface-hub==0.23.4
idna==3.7
inflect==7.2.1
influxdb==5.3.2
influxdb-client==1.48.0
intel-openmp==2021.4.0
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
jsonpickle==4.0.1
kiwisolver==1.4.5
konlpy==0.6.0
Korpora==0.2.0
langdetect==1.0.9
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mecab-python3==1.0.9
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opcua==0.98.13
openpyxl==3.1.2
packaging==24.2
pandas==2.2.1
parso==0.8.4
pefile==2023.2.7
pillow==10.3.0
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pyecma376-2==1.0.0
Pygments==2.19.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pymongo==4.6.3
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-engineio==4.11.2
python-pptx==0.6.23
python-socketio==5.12.1
pytz==2024.1
pyvis==0.3.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rdflib==7.1.1
reactivex==4.0.4
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
simple-websocket==1.1.0
six==1.17.0
soynlp==0.0.493
stack-data==0.6.3
sympy==1.13.0
tbb==2021.13.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.3
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
xmltodict==0.14.2
